empty commit with notes about merging repos
This used to be a debian branch on genome-vendor/pindel0.2.4t repo.  I
moved it over to genome/pindel and setup gbp.conf to work with the new
naming scheme.

One of the commits in genome/pindel's master is erroneously tagged as
v0.2.4t so I had to change the upstream-version scheme to be
"r%(version)s" instead of the default "v%(version)s".  We should
probably switch back to the default on next release.

The commit tagged r0.2.4t was identical to genome-vendor/pindel0.2.4t's
upstream branch.